+++ front/package.json
@@ -15,11 +15,13 @@
     "js-cookie": "^3.0.1",
     "less": "^3.0.0",
     "less-loader": "^6.0.0",
+    "quill": "^1.3.7",
     "register-service-worker": "^1.7.1",
     "sass": "^1.39.0",
     "view-design": "^4.0.0",
     "vue": "^2.6.11",
     "vue-i18n": "^8.25.0",
+    "vue-quill-editor": "^3.0.6",
     "vue-router": "^3.2.0",
     "vuex": "^3.4.0"
   },
