+++ front/src/components/Empty.vue
@@ -1,0 +1,15 @@
+<template>
+  <div>
+    404 不存在
+  </div>
+</template>
+
+<script>
+export default {
+  name: "Empty"
+}
+</script>
+
+<style scoped>
+
+</style>
