+++ front/src/main.js
@@ -7,9 +7,14 @@ import store from './store'
 import i18n from './i18n'
 import less from 'less'
 import ViewUI from 'view-design';
+import VueQuillEditor from 'vue-quill-editor'
 
 // css
 import './theme/bfban_theme.less'
+// quill css
+import 'quill/dist/quill.core.css'
+import 'quill/dist/quill.snow.css'
+import 'quill/dist/quill.bubble.css'
 
 // js
 import './registerServiceWorker'
@@ -17,6 +22,7 @@ import './plugins/iview.js'
 
 Vue.use(less)
 Vue.use(ViewUI);
+Vue.use(VueQuillEditor, /* { default global options } */)
 
 Vue.config.productionTip = false
