+++ front/src/theme/bfban_theme.less
@@ -34,7 +34,6 @@
     cursor: pointer;
     border: 1px solid @border-color-base;
     border-left: 0;
-    background: #fff;
     position: relative;
 
     > span {
