+++ front/src/views/Reset.vue
@@ -1,27 +1,33 @@
 <template>
   <div class="container">
     <div class="content">
-      <Form :label-width="80" style="position: relative;">
-        <Divider>{{ $t("reset.title") }}</Divider>
+      <Breadcrumb>
+        <BreadcrumbItem to="/">{{ $t("header.index") }}</BreadcrumbItem>
+        <BreadcrumbItem>{{ $t("reset.title") }}</BreadcrumbItem>
+      </Breadcrumb>
+      <br>
 
-        <FormItem :label="$t('reset.form.qq')">
-          <Input type="text" v-model="reset.qq" :placeholder="$t('reset.form.qq')"/>
-        </FormItem>
+      <Card dis-hover>
+        <Form :label-width="80">
+          <FormItem :label="$t('reset.form.qq')">
+            <Input type="text" v-model="reset.qq" :placeholder="$t('reset.form.qq')"/>
+          </FormItem>
 
-        <FormItem :label="$t('reset.form.password')">
-          <Input type="password" v-model="reset.password" :placeholder="$t('reset.form.password')"/>
-        </FormItem>
+          <FormItem :label="$t('reset.form.password')">
+            <Input type="password" v-model="reset.password" :placeholder="$t('reset.form.password')"/>
+          </FormItem>
 
-        <FormItem :label="$t('reset.form.passwordRepeat')">
-          <Input type="password" v-model="reset.passwordRepeat" :placeholder="$t('reset.form.passwordRepeat')"/>
-        </FormItem>
+          <FormItem :label="$t('reset.form.passwordRepeat')">
+            <Input type="password" v-model="reset.passwordRepeat" :placeholder="$t('reset.form.passwordRepeat')"/>
+          </FormItem>
 
-        <FormItem>
-          <Button @click.prevent.stop="handleReset" type="primary">{{ $t('reset.form.submit') }}</Button>
-        </FormItem>
+          <FormItem>
+            <Button @click.prevent.stop="handleReset" type="primary">{{ $t('reset.form.submit') }}</Button>
+          </FormItem>
 
-        <Spin size="large" fix v-show="spinShow"></Spin>
-      </Form>
+          <Spin size="large" fix v-show="spinShow"></Spin>
+        </Form>
+      </Card>
     </div>
   </div>
 
@@ -82,4 +88,3 @@ export default {
 
 <style>
 </style>
-
